test/voice_automation/requirements_voice_test.txt
# Voice AI Test Automation Dependencies
# Install with: pip install -r requirements_voice_test.txt

# Robot Framework Core
robotframework>=7.0
robotframework-browser>=18.0
playwright>=1.41.0

# LiveKit SDK
livekit>=0.12.0
livekit-api>=0.6.0

# WebSocket Support
websockets>=12.0

# HTTP Client
aiohttp>=3.9.0
httpx>=0.27.0

# Audio Processing
numpy>=1.24.0
soundfile>=0.12.0
scipy>=1.11.0

# Utilities
python-dotenv>=1.0.0
dataclasses-json>=0.6.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
